Refactor image functions into util file
